STRUC-440: Adding the advanced modeling tab and the first advanced modeling setting

+++ apps/structures/app/_job-based-range/shared/effects/shared.effects.ts
@@ -3,7 +3,7 @@ import { Injectable } from '@angular/core';
 import { Actions, Effect, ofType } from '@ngrx/effects';
 import { Action, select, Store } from '@ngrx/store';
 import { Observable, of } from 'rxjs';
-import { switchMap, map, mergeMap, catchError, withLatestFrom } from 'rxjs/operators';
+import { switchMap, map, mergeMap, catchError, withLatestFrom, tap } from 'rxjs/operators';
 
 import { StructureModelingApiService } from 'libs/data/payfactors-api/structures';
 import * as pfDataGridActions from 'libs/features/pf-data-grid/actions';
@@ -13,13 +13,13 @@ import { NotificationLevel, NotificationSource, NotificationType } from 'libs/fe
 import * as fromPfDataGridReducer from 'libs/features/pf-data-grid/reducers';
 import { DataGridToDataViewsHelper } from 'libs/features/pf-data-grid/helpers';
 import { DataViewApiService } from 'libs/data/payfactors-api/reports';
+import * as fromRangeFieldActions from 'libs/features/structures/range-editor/actions/range-field-edit.actions';
 
 import * as fromSharedActions from '../actions/shared.actions';
 import { PayfactorsApiModelMapper } from '../helpers/payfactors-api-model-mapper';
 import { RangeGroupMetadata } from '../models';
 import * as fromSharedReducer from '../reducers';
 import { PagesHelper } from '../helpers/pages.helper';
-
 
 @Injectable()
 export class SharedEffects {
@@ -75,17 +75,28 @@ export class SharedEffects {
       ofType(fromSharedActions.GET_OVERRIDDEN_RANGES),
       switchMap(
         (action: fromSharedActions.GetOverriddenRanges) =>
-          this.structureModelingApiService.getOverriddenRangeIds(action.payload.rangeGroupId)
+          this.structureModelingApiService.getOverriddenRanges(action.payload.rangeGroupId)
             .pipe(
               mergeMap((response) =>
                 [
-                  new fromSharedActions.GetOverriddenRangesSuccess(),
-                  new fromPfDataGridActions.UpdateModifiedKeys(action.payload.pageViewId, response)
+                  new fromSharedActions.GetOverriddenRangesSuccess(response),
+                  new fromPfDataGridActions.UpdateModifiedKeys(action.payload.pageViewId, response.map(o => o.CompanyStructuresRangesId))
                 ]),
               catchError(error => of(new fromSharedActions.GetOverriddenRangesError(error)))
             )
       )
     );
+
+  @Effect()
+  refreshOverriddenRanges$: Observable<Action> = this.actions$
+    .pipe(
+      ofType<fromRangeFieldActions.UpdateRangeFieldSuccess>(fromRangeFieldActions.UPDATE_RANGE_FIELD_SUCCESS),
+      map(action => {
+        return new fromSharedActions.UpdateOverrides({ rangeId: action.payload.modifiedKey, overrideToUpdate: action.payload.override});
+      })
+    );
+
+
 
   @Effect()
   revertingChangesRange$: Observable<Action> = this.actions$
@@ -98,10 +109,14 @@ export class SharedEffects {
             action.payload.rangeGroupId,
             action.payload.roundingSettings))
             .pipe(
-              mergeMap(() => {
+              mergeMap((response) => {
                 const actions = [];
-
-                actions.push(new fromPfDataGridActions.DeleteModifiedKey(action.payload.pageViewId, action.payload.rangeId));
+                // only remove the key if the override was deleted
+                if (response.OverrideDeleted) {
+                  actions.push(new fromPfDataGridActions.DeleteModifiedKey(action.payload.pageViewId, action.payload.rangeId));
+                } else {
+                  actions.push(new fromSharedActions.UpdateOverrides( { rangeId: action.payload.rangeId, overrideToUpdate: response.Override }));
+                }
                 actions.push(new fromSharedActions.RevertingRangeChangesSuccess({
                   pageViewId: action.payload.pageViewId,
                   refreshRowDataViewFilter: action.payload.refreshRowDataViewFilter,
